Business Analysis Using Market Basket Analysis
This part of this study is aimed at performing the Market Basket Analysis also called MBA. There are many data analysis tools available to the python analyst and it can be challenging to know which ones to use in a particular situation. A useful (but somewhat overlooked) technique is the association analysis which attempts to find common patterns of items in large datasets. For this section the online-retail dataset will be used to analysis the Market Basket Analysis also known as MBA, using the Association rules for better understanding of the dataset.  The one we will be looking at is the Market Basket Analysis. The very useful MLxtend library by Sebastain Raschka has an implementation of the Apriori algorithm for extracting frequent item sets for further analysis.
Why Association Analysis?
Association Analysis is relatively light on the math concepts and easy to explain to non-technical people. In addition, it is an unsupervised learning tool that looks for hidden patterns so there is limited need for data prep and feature engineering. It is a good start for certain cases of data exploration and can point the way for a deeper dive into the data analysis using other approaches. 
 Product recommendation algorithms account for 35% of all Amazon purchases. (McKinsey, 2013). Recommendation systems have an impact on our daily lives and basic form of recommendation system we will be looking at is Market Basket Analysis. Market Basket Analysis is a simple concept that does not necessitate advanced statistics or mathematics understanding. (Frank La, 2018). 
In Market Basket Analysis attributes like, consequent, support, confidence, lift is used for evaluation. Where support helps to study the relative frequency that the rules show up. A high support helps to show relationship. However, an instance where a low support is useful if you are trying to find “hidden” relationship. For the confidence, for product recommendation, a 50% confidence may be perfectly acceptable but in a medical situation this level may not be high enough. Lift is the ratio of the observed support to that expected if the two rules were independent. The basic rule of thumb is that a lift value close to 1 means the rules were completely independent. Lift > 1 are generally more “interesting” and could be indicative of a useful rule pattern. 
2.Application
Market Basket Analysis is used to understand the purchasing behavior of customers. It leads to effective sales and marketing. It can be used for store layout, marketing messages, maintain inventory, content placement and recommendation engines. Its application cut across affinity promotion, product placement, cross selling, fraud detection and customer behavior. Under the MBA association rule, this technique aims at searching data for frequent if-then patterns and by using a certain criterion under support and confidence to define what the most important relationships are. Support is the evidence of how frequent an item appears in the data given, as confidence is defined by how many times the if-then statements are found true. There is a third criteria called the lift and this can be used to compare the expected confidence and the actual confidence. The lift shows how many times the if-then statement is expected to be found to be true. Association rules are calculated from itemsets, which are created by two or more items. If the rules were built from the analyzing from all the possible itemsets from the dataset. 
Apriori Algorithm:
The algorithm is for frequent item set mining and association rule learning. It proceeds by identifying the frequent individual items in the database and extending them to larger and larger item sets as long as those item sets appear sufficiently often. It uses prior knowledge of frequent itemset properties.